Module pulsar::commands::farm

Constants

BATCH_BLOCKS ğŸ”’

N_TASKS ğŸ”’

SINGLE_INSTANCE ğŸ”’

allows us to detect multiple instances of the farmer and act on it

Functions

farm ğŸ”’

implementation of the farm command

get_author_info_from_blocks ğŸ”’

not_yet_processed_block_nums_stream ğŸ”’

plotting_progress_bar ğŸ”’

nice looking progress bar for the initial plotting :)

process_block_stream ğŸ”’

subscribe_to_node_syncing ğŸ”’

subscribe_to_plotting_progress ğŸ”’

subscribe_to_solutions ğŸ”’

syncing_progress_bar ğŸ”’

nice looking progress bar for the syncing :)

wait_on_farmer ğŸ”’

Type Aliases

MaybeHandles ğŸ”’