Function pulsar::utils::get_user_input

pub(crate) fn get_user_input<F, O, E>(
    prompt: &str,
    default_value: Option<O>,
    condition: F
) -> Result<O>where
    E: Display,
    F: Fn(&str) -> Result<O, E>,

gets the input from the user for a given prompt

default_value: will be used if user does not provide any input

condition: will be checked against for the user input, the user will be repeatedly prompted to provide a valid input

error_msg: will be displayed if user enters an input which does not satisfy the condition